+++ aws/oidc/jenkins-oidc/main.tf
@@ -4,4 +4,5 @@ module "provider" {
   url                    = var.url
   client_id_list         = var.client_id_list
   additional_thumbprints = var.additional_thumbprints
+  create_provider        = var.create_provider
 }