+++ aws/oidc/provider/CHANGELOG.md
@@ -4,4 +4,4 @@ All notable changes to this project will be documented in this file.
 
 ## [0.2.0]() (2024-12-24)
 ### Features
-* Init and generalize module `provider` to [(aws/oidc/provider)](../provider).
+* Move and generalize module `provider` to [(aws/oidc/provider)](../provider).